test(monatsplaner): fix mocking issues cross suite
There are some awkward effects going on when mocking whole directories.
Mocked modules are somehow not properly reset when doing so. After a lot
of try-and-error, the best working solution is to mock individual
modules only. This is a little annoying in this specific case. But just
shows the number of dependencies this piece of code has.