refactor(monatsplaner): move gueltiger Plan specification out of Plan data structure
The past concept was to provide the specification for a gültigen Plan as
property of the Plan data structure. The thoughts behind this concept
where primarily driven by better testability.
But this concept does not scale as we develop the need for multiple
specifications for various purposes. Keeping them part of the data
structure does not seem to match anymore.
The actual testability part is only needed for basically one single
test "module". Which can be simply resolved by proper mocking. As usual
for unit tests. The other related test suite could have used proper
mocking in first place.

This resolves the need to provide a specification anywhere a Plan is
needed as test data. Because there are still no proper data factories,
this was still provided everywhere. In result, the Top and Bottom
specifications do no more serve a proper need and could have been
removed.