feat(monatsplaner): add specification for final validation
The naming to clearly communicate the purpose is not yet on point. But
the original specification for a gültigen Plan has been "reduced" to an
intermediate level. The new final specification builds on top of it an
validates everything. Including rules which can't be applied while
planning.

The new operation to validate a plan for the final submission just wraps
the evaluation of the specification to make "better" usable from outside
the domain. It is currently under discussion how exactly this will be
triggered.