feat(monatsplaner): make disabled Optionen focusable
In a radio-button group, any disabled option can not be focused. This is
especially confusing for usage screen readers.
Assuming Basiselterngeld is disabled. Before this, the user would tab
into the fieldset with the radio-button group and immediately lands on
the second option (ElterngeldPlus). But the screen-reader will announce
it as 2nd of 4 options. Cycling through the options with the arrow keys
will jump from 2 to 3 to 4 to 2 again.
With this improvements, disabled radio buttons are still focusable, but
are announced as dimmed/disabled by the narrator. Trying to select them
does nothing. For keyboard usage this means that the last option remains
chosen till hitting the arrow key again.