feat: improve a11y of info dialog component
This mostly just implements best practices for such dialogs based on
accessibility guidelines. In practice it actually helps a lot when using
a screen reader.

The improvement was especially tested for the Auswahl of the Optionen in
the Planer. There it also got improved by the exposed label property
now.

It also lead to some simplifications of the info dialog component which
lead to slightly broader changes. Though, it is not the goal right now
to fully rework the info dialog from the bottom up. Which is planned for
a later point in time. Because it has many more issues.