style(monatsplaner): add focus styling for Auswahl label
The Auswahl Eingabe uses a radio button group. But visually they appear
like buttons. This is achieved by hiding the input themselves and style
the labels properly. But this does properly communicate they focus state
as this is hidden too. Thereby, it was really hard to follow the focus.

The final styling of the outline is still under review.