test(monatsplaner): add test cases for Mutterschutz
During manual operation it has shown that Mutterschutz seams sometimes
special. Just to be absolutely sure, but also for the communication of
the tests as documentation these test cases have been added.

Minor additional change for test factory functions to not use optional
parameters but ones that can be undefined. That requires the caller to
explicitly set an undefined value which should improve readability.