fix: control focus restoration of info dialog
When improving the accessibility of the info dialog component, the focus
management was to "aggressive".
The guidelines for such dialogs dictate to recover the focus after
closing. But this was also applied for the "click-outside-action". This
caused a lot of issues for many interactions everywhere. In general was
this event was triggered way to often even when not relevant.

It also includes a fix for a typo.

For the protocol:
This is a late fix because I did not run the full test suite and the CI
is currently not "connected" for this feature branch.