feat(monatsplaner): improve screen reader support for Auswahl Eingabe
Due to the slim layout and to keep things more terse, the design is not
very verbose. For example does "Basiselterngeld" gets shorten to just
"Basis". And in general does the existing text not read well when
composed by the raw accessibility DOM.
This is not optimal and not too nice in code. But to provide a better
user experience for those who need it, this is very helpful.

Moreover, the linking to the hints why an Option is disabled gets better
communicated. Unfortunately there is no optimal way to do it. Because
the user is "trapped" within the radio-button group and the related info
dialogs are outside of it and only visually linked. The probably best
approach is to use the `aria-details` properly. In theory it should be
announced to the user with its label (of the referred details) and
allows the user to jump to the details (in our case the button to open
the info dialog). But the support of this attribute is extremely pure
and insufficient.
That's why some additions are needed. First of all, the info dialogs are
in the tab order anyways. Therefore, they come before the radio-button
group. Because they are part of the fieldset, they get announced in the
right context. They give the user an "initial" overview why things are
available or not. But he is free to actually open the dialog or just hop
over them. Additionally the hint gets read out as description of each
disabled input. As description, it comes quite last in the announcement.
So it is not too disturbing and can be skipped.