style(monatsplaner): remove icon for details collapsed state
For the Lebensmonate Details horizontal space is very critical. The icon
which indicates if the details are closed or open take valuable space.
Furthermore, it is asymmetric on the right side only which causes more
challenges.
We decided to remove the icon(s). We assume that with all the other
indicators (hover effects, plus icon, ...) users know how to expand and
collapse the Lebensmonate. We need to test/observe this.