feat(monatsplaner): use more clear sections with headings
The advantage of using headings to label the sections instead of a plain
label attribute is the prominence. It stronger communicates the
structure and sections of the applications in multiple screen-reader
wheels. It especially allows users to interact with the planer in more
advanced patterns when trying multiple plans etc.