chore(monatsplaner): fix typo and export operation
First of all, many of the changes in this commit address a stupid typo
of a duplicated letter.

Second, this exports the operation to split Lebensmonate by Elternteile
into an own operation. This is with the anticipation of re-usability,
as future features will need it soon (just trying to keep commits
smallish).