feat(monatsplaner): migrate components for Plangunsdetails
Also it might not look like it on a first gaze, these components have
been directly copied over. Though, they went under a view changes. Most
of them are pure adaptions to the new domain and its naming. This also
includes the logic to trim and fill gaps in the planned Lebensmonate.

It also removes the whole income part for now. The integration of the
Rechner into the Planer is still standing out. But more importantly we
learned that these income values can't be used straight. We have to find
an approach that works fine.

There were no tests in the past and currently none have been
supplemented yet.