feat: add new hook to navigate and pass on a Plan
This is an alternative and cheap approach to pass the planned Plan
between the Planer page and the Zusammenfassung. It simply uses the
browsers history state. It is a replacement for a state. There is no
real state management necessary for this part yet. Only the Planer edits
it and the rest just reads/views the Plan.
That means it might get replaced with a more advanced solution once
needed.

Defining this behavior as a hook helps with the unified "encoding" and
usage. It also makes it easily testable in a decoupled manner.

Tests are split into a specification file instead of in-source tests,
because the latter cause global mocking issues right now.