refactor(monatsplaner): improve Ausgangslage composition
The store selector was simply overkill. Especially because it had the
setup to never update and only execute once on first render. In result
a simple function which takes the stores current state is fine too. But
it is more simple and easier to test.
Furthermore, once a Plan was created using the Ausgangslage, the latter
is embedded into the first and should only used this way. This avoids
any confusions or out-of-sync issues.