feat(monatsplaner): remove visible label for Lebensmonatsliste
It was discussed and argued that this is confused to be either a heading
or a column label (table header). And in turn, if it would be recognized
as column label, why are the other columns not labelled. In the end it
was decided to remove it.

For non visual users it still remains to act as label and landmark for
navigation and overview.