refactor(monatsplaner): split Lebensmonat Details component
This was from the beginning one of the most complex components and had
to break down at some point. Especially knowing about future features
like Einkommen.
In result, the individual components have all a clear purpose. Though,
there pop drilling continues and the directory tree gets deeper. The
letter just relates to the DOM structure as it is. And the props are to
the majority view related readonly data.

This also introduces the new "concept" of managing the grid layout down
the tree. It remains of the main responsibilities of the Details
component to define the grid template and areas. This will be used even
more soon.

Notice that the naming scheme of the components starts to get some HTML
technical aspect. Like the big Lebensmonat gets split into the details
and its children summary and content.