style(monatsplaner): drill down nested grid for Lebensmonat
After many discussions with design, the grid layout is still
problematic. Especially the shortcuts down the component tree does not
work out to well. Therefore, to make the layout work nicely, also for
single Elternteil, a deep sub-grid is needed.

The strain on the components and the entangling is big. Components do no
more just receive some "anonymous" `className` properties and only the
top level component cares about the grid. I tried to make the structure
well enough and consistent that it is hopefully readable and
understandable.
The grid for the single Elternteil has improved completely based on
design input.