fix(monatsplaner): subgrid within fieldset
The new deep subgrid was not working in Chrome (i.e. the Blink engine).
The issues was that subgrids do not work within fieldset elements.
Apparently there is a specification issue or similar.
The currently recommended workaround is to use `display: contents`
styling instead of a subgrid. But as we still need a subgrid for other
styling purposes, a wrapper was added. In result we use a subgrid and
a display-contents together.

Moreover there was a simplification in terms of grid rows because the
subgrid is limited to the columns only.