chore(monatsplaner): remove old code due to new domain
This gets rid of the whole `monatsplaner` package (was formerly
a package). This includes the related Redux store and many related
things.

The `ElterngeldType` was used excessively among the whole form-page part
of the application. So it was extracted for now as a placeholder.

The `EltergeldVarianteBadge` was migrated into the only component it was
used for simplicity and type issues.