feat(monatsplaner): improve texts and labes for single Elternteile
A single Elternteil has no Pseudonym anymore. This also fixes a "bug"
when switching between `für beide` and `für einen Elternteil`. In result
this makes the pseudonym also become "hidden" everywhere, like design
wants to without further ado.

Moreover, too many ARIA labels and descriptions made use of the
pseudonyms. This was relying on the default of `Elternteil 1` (and or
`Elternteil 2`) for the pseudonyms. But it is a bit awkward to use this
term if there is only one Elternteil at all. So texts have been adapted
to use pseudonyms only where appropriate.