test(e2e): update for usage and screenshots of new planer
The end-to-end tests are in an awkward intermediate state where some are
using the POMs and others not. The latter have been quickly added with the
Playwright recording tool to cover production better.
As intermediate solution, the POM got extended to handle the selection
within the Planer. All auto-generated test code which does not relate to
this remained untouched for now.

Screenshots where kinda just accepted with a short glimpse. Things have
changed and mismatches are expected. But the new baseline is set now and
will help with regression testing.