feat(monatsplaner): use preconditions for Plan validation
Uses the new feature of specifications to optimize the user experience.

This refactors the specification for the Kontingent to make it usable
for preconditions. In result the new implementation is actually more
simple and does not need a class.
Furthermore, this should allow to optimize the Basis vs. Plus Kontingent
issue in future.