style(monatsplaner): align Kopfleiste with Lebensmonaten
The width of the Lebensmonate adapts with the number of Elternteile.
That's why it was included in the grid definitions, which are also
defined by number of Elternteile.
But because of the shadow DOM restrictions of the `<details>` elements,
the grid start only within these. That means a grid alignment with the
Kopfleiste is not easily possible. Anyhow, the usage of flex
justification gave pretty good results. But only if the Kopfleiste
matches the width of the Lebensmonate.
Therefore, the width styling is done on a higher level now to make this
possible. The whole orchestration of varying widths in the overall
Planer component is not optimal. Especially the block background colors
make it a little complicated. But the code should be still good
maintainable for now.