style(monatsplaner): fix responsive layout
With the last attempt to properly place the Pseudonyme der Elternteile
in the Kopfleiste above the Lebensmonate just using flexbox, the
responsive layout was broken.
In general it didn't feel great. After all, using the grid through out
seems to be the only way to have a proper alignment. Though, the shadow
DOM of the `<detail />` elements still causes issues.
The resulted implementation is not great. But in general is a huge grid
with subgrids not optimal for component tree independence. The current
approach should be fine for now. Though, we should keep an eye on it.