fix(monatsplaner): responsive view for gewählte Option
Somehow, for unknown reasons, some container query utility classes of
TailwindCSS do not work. For example was the estimated width for Monate
mit Mutterschutz `14ch` and for `hinzfügen` it was `12ch`. The class
`@[12ch]:block` did work, but `@[14ch]:block` did not. This was quite
often the case. Every time I tried to optimize the width estimation
function it might not work at all anymore.
Therefore I decided to go with a more stable/fixed class that is known
to work. The currently only two cases work pretty well with the same
value. Future cases are not in sight and unlikely. Anyhow, future
problem.