refactor(monatsplaner): try to get deep grid under control
The deep grid makes some issues in terms of code orchestration. The grid
is deep on multiple levels. First of all, it uses `grid-column-template:
subgrid` and `display: contents` in the actual element tree of the
browsers DOM. Furthermore, it spreads cross the React component tree in
the code.

The prop-drilling and grid layout type definitions do not scale and
provide a horrible developer experience. Therefore the grid has been
more decentralized.
Therefore, each component who wants to place items can define this
itself independently. No parent component needs to support this in any
way. Except for the grid layout to be applied at any higher level.

To achieve this, the whole grid has been transformed into TypeScript
code and gets attached to the components via CSS styling properties. No
intermediate utility classes or CSS modules.
This allows to have a strong relationship between the grid template
definitions and the components placing their elements based on grid
columns. The result is kind of an open-closed concept. It scales really
well, the code units are loosely coupled via "interfaces" and the
developer experience is pretty great.

Let's see how this ages.