fix(monatsplaner): disable click-outside feature for touch on Lebensmonate
This causes the weird bug that first the open Lebensmonat closes and
then the tap triggers the element which is now visible at the location
after the closing. Example: Lebensmonat 1 is open. The user taps below
on Lebensmonat 2. Lebensmonat 1 closes and where was Lebensmonat 2 is
now Lebensmonat 5 in the collapsed view. Lebensmonat 5 opens instead.

This does not seem to be any issue on desktop with click events in any
tested web-browser.