fix(monatsplaner): keep Bruttoeinkommen when switching from Bonus
There was the bug with the Bruttoeinkommen. The scenario is
a Lebensmonat with Bonus chosen and some Bruttoeinkommen specified. When
switching now from Bonus to another Option, the Bruttoeinkommen was
deleted.
This was due to the `setzeMonatZrueck` operation and its siblings. With
the addition of the Bruttoeinkommen, the operation is no more accurate.
Actually it should only reset the chosen Option and the related
Elterngeldbezug. To fix that and to properly express the operations
purpose again, it got also renamed.