feat(monatsplaner): add specification for Bonus and one Elternteil
While already covered with the zero verfügbares Kontingent, this new
specification allows to improve the user experience a little.
Furthermore this makes the rule more "prominent" and not "hidden" in the
Kontingent.
The usage of specification pre-conditions starts to get a little wild to
get the best user experience and have the rules actually be applied. But
so far it still looks okay.