refactor(monatsplaner): reduce prop drilling with context
The component tree for a singe Lebensmonat has become quite wide and
deep. Many of the components need access to the same information. Due to
that there is some prop-drilling going on. That was acceptable for
a while, but reached a peak now.
To resolve this issue, the probably most obvious solution by React was
used: a context. Simple, straight forward, good enough for now.