fix(monatsplaner): performance issues in Lebensmonatsliste
Since the addition of the input fields to specify a Bruttoeinkommen, the
application suffers in performance. This is because every single
Lebensmonat and its component does re-render on any change of the Plan
data. That is due to the `bestimmeAuswahlmoeglichkeiten` callback must
be tied to the Plan state. So every time the Plan state is set, the
callbacks change.
Theoretically there are many more properties down the component tree
that cause re-render due to being dependant on the Plan state. But this
callback is a critical one, because it can't be fixed with any kind of
memorization.

The actual solution is to only render the content of a Lebensmonat when
this Lebensmonat is expanded. In result, also the Auswahlmöglichkeiten
are always requested for the open Lebensmonat(e).