refactor(monatsplaner): improve dependencies of Planer service
This commit has become a little big and should have been split into 2 or
3. Unfortunately was development already too far and splitting it again
was too involved/time costly.

First and for most, the dependency of the planer feature to the rest of
the application (i.e. the Redux stores) was reversed. Thereby the planer
is an independent feature now, which at most depends on a tiny set of
atom components.
In practice this means that the code that uses the Planer must provide
either an old plan to continue planning from (as optionally before) or
an Ausgangslage to start a new plan from.

Furthermore, this provides some minor "performance" optimization. It
does so by skipping certain state updates (and thereby calculations) in
the Planer when specifying a Bruttoeinkommen. Like it is not necessary
to update the verplantes Kontingent. This also helps to reduce the
need for re-rendering.

Finally, the tests of the Planer service have been restructured and
improved a little. The order of testing follow more along with the
source code. The service code is quite big by now and their unit tests
too. Help with readability is key here.