refactor(calculations): make calculation synchronous
The only reason the whole calculation was asynchronous was the
possibility to calculate the taxes via a remote service. This service
was no more used. We don't event have an authentication token anymore.
Moreover is this external service about to become obsolete.
After all, the asynchronous part is causing issues. At least when it can
synchronous. It makes so many things more simple. There is no
intermediate calculation state or anything. All straight forward.

The current usage with asynchronous thunks in the Redux store remains as
it. It does not break technically. And this store gets removed soon
anyway. Therefore no further improvements down the pipe for now.