feat: add hook to provide calculation of Elterngeldbezuege
This is some kind of a "preparation commit". It encapsulate this feature
to keep the commit size reasonable. Though, it goes hand in hand with
the follow up commit which will make use of the hook.
It provides a new entrypoint that connects the data from the Redux store
and triggers the Eltergeldrechner to calculate Elterngeldbezuege. This
entrypoint can be used very flexible without knowing the internals or
a direct dependency to the store.