feat(monatsplaner): integrate with Elterngeldbezug calculation
This makes the Bruttoeinkommen being used to influence the calculation
of the Elterngeldbezuege. It therefore uses a callback function as
dependency inversion/injection approach. The calculation should stay
outside of the domain or rather the planer feature as whole.

There was a chicken or egg problem with the creation of a new Plan.
The data structure of a plan includes the last calculated
Elterngeldbezüge. But to calculate the initial Elterngeldbezüge, the
initial Lebensmonate are needed. Thereby, the operation to create an
initial Plan was removed and the object gets simply assembled directly,
creating the initial Lebensmonate first.