fix(monatsplaner): toggle Einkommen closes Lebensmonat
There was the bug with the toggle button that opens and closes the
section with the Einkommen of a Lebensmonat. When clicking the icon in
the button, the whole Lebensmonat was closed.

The issue was due to the click-outside feature of the Lebensmonat that
closes it. Though, the icon of the button is within the Lebensmonat. But
the icon is changing depending on the toggle state. So when the `click`
event is handled, the target element of the click (the old icon) is no
more in the DOM, especially not within the Lebensmonat element. Thereby,
this is detected as an outside click. React is basically too fast.
Using the `mousedown` event solves this issue. But it brings new issues.
The toggle works fine. Clicking on the next closed Lebensmonat does no
more work.

Therefore, the solution is little hacky. I simply offsets the state
change of the toggle just enough to make the click event handled first.