feat: use custom hook for click-outside effect
While debugging the last issue, the `react-detect-click-outside`
dependency needed to be inspected. Turned out this is a trival
implementation. For a period of time it looked like a custom version is
necessary to fix the bug. In the it was not. But seems no reasonable to
maintain a dependency for such a trivial thing. It's basically just
"regular code". It is more tailored to our needs now. Plus it marginally
reduces the final bundle size.