feat(monatsplaner): prevent Bruttoeinkommen im Monat mit Mutterschutz
In a Monat mit Mutterschutz it is not possible to have eine Einkommen.
Therefore the input for the Bruttoeinkommen should not be available for
the respective Elternteil for these Lebensmonate.

This has an ugly twist on the layout. Because the hint of for the
number of allowed working hours should be displayed differently
depending on the number of inputs for Bruttoeinkommen. If there are not
much further exceptions, the component is managable.