refactor: break dependencies between planer and tracking
This was kind of the worst scenario: the planer feature had a dependency
on the tracking (directly) and the tracking on the planer feature.
Instead the dependencies should be reversed for both cases to create
independence and more stable separate features.
In the end there must be always a "dirty" part which links the parts
together. Which in this case is the page view component.

Notice that we still have no test concept for the page view components.
So the integration remains untested for now.