chore(hooks): use --no-warn-ignored flag instead of glob pattern
With the adoption of the flat config, we can now leverage the
--no-warn-ignored flag to suppress warnings for ignored
files. This change allows us to exclude files from linting
rather than specifying a whitelist of files to lint. It also
ensures that eslint.config.js remains the single source of
truth for determining which files should be linted.