test: add end-to-end test case for accessibility
This is a first effort into this direction. Further (conceptional)
improvements are planned, but not yet fully designed.

This includes a bunch of smaller fixes to address discovered
accessibility violations.
Many were related to correct heading levels. Most notably was the issue
of attaching an `aria-details` attribute to each Auswahloption, even for
those which have no info dialog.

There is currently a final violation which could have not been easily
resolved. That is the info dialog within a details summary element. This
requires some conceptional changes to the UI. This is ignored for now.