test: update and fix end-to-end tests
Context:
We had a long living feature branch with breaking changes. To resolve
this branch, an interactive rebase was chosen to get in control of all
the conflicts, updated linting etc.
Anyhow, on this feature branch the end-to-end tests were not maintained.
Now the tests have caught up again. Also the screenshots have been
updated.

Notice that the originally fully generated end-to-end tests slowly
transition into hand written once. We need to evaluate how to continue
with this.