test: improve a11y test configuration
According to the documentation of Axe, this should increase the
performance slightly. Anyhow, as we really never access anything else of
the Axe resport but the violations, this is just a fine addition.
It is hard to gather some statistics how much this actually saves as, as
the test execution duration varies a lot. The hope is, that with further
usage of a11y testing, this will save us some time in the long run.