feat(monatsplaner): add Vorschläge for Angabe des Einkommens
This feature as a first effort to make the input of a Bruttoeinkommen
over and over more simple. It does so by providing suggestions for the
input based on the past inputs. The current suggestion "algorithm" is
rather simple.

There are many files touched here again. Partially due to the
prop-drilling (which still exists). But also due to the context
provision across test cases.