style: unify spelling for Bruttoeinkommen
There was a mix of spellings for this word. Especially in old code the
`Brutto-Einkommen` version was used. According to the Familienportal it
should be `Bruttoeinkommen`.

Notice that many old code sections still use the camel case form
`bruttoEinkommen`. In contrast to the aligned `bruttoeinkommen` in more
recent code. This hasn't been fixed yet as it has no real benefit.