feat(monatsplaner): improve order for Bruttoeinkommen Vorschläge
The suggestions should help the user to make input quickly and
efficiently. The likelihood that following Lebensmonate have the similar
Bruttoeinkommen is higher than any other. Like for the Bruttoeinkommen
of the 12th Lebensmonat, getting a suggestion from the input done at the
11th is probably more useful than from the 2nd.
By the nature of the first simple algorithm, the suggestions where
sorted by the Lebensmonate. Therefore the order is basically exactly
reversed to the above principle.
Moreover, if the user starts planning linearly and making adjustments to
any "random" Lebensmonat, a simple sorting does not work out anymore.
Therefore, Vorschläge are sorted by being closest to the current
Lebensmonat they are suggested for. This remains stable cross removing
duplicated entries.