fix(routing): use native scroll restoration
The effect hook to scroll to the top after each navigation did not
reliably work. It could be argued that this is no idiomatic React. But
there is some native support this feature by the routing library itself.
Though, it does no more use smooth scrolling. Which isn't the case for
regular routing anyhow. If we don't wanna use routes anymore, we might
revisit this decision in general.

This required to adapt the test utilities to support component tests
which use this component (only works with data routers).