chore: fix code smell for aria log context
The original code smell was too deeply nested functions and thereby too
high complexity.
While working on it, the code wasn't really clean. The add message
function which returns a remove message function so a React effect is
cleaning up when a component get unmounted is not straight forward.
The solution approach is to break down the code by separating these
concerns. In result the nesting is resolved automatically.