chore: road to zero code smells
We're trying to use our code quality gate more strict. To do this
efficiently, it is best to set the number of allowed code smells to
zero. Else it is an endless battle.

This resolves most of the known issues for the `src/globals` directory.
It were primarily issues with the critical calculation with is
absolutely horrible code and remains untouched. Therefore, all fixes
happened in the most simple and stupid way possible.
Unfortunately there are issues which can't be resolved in KISS way.
First and for most this are complexity issues.