chore: road to zero code smells
We're trying to use our code quality gate more strict. To do this
efficiently, it is best to set the number of allowed code smells to
zero. Else it is an endless battle.

Disable a rule that claims that many `as const` assertions are not
necessary. But in fact they are to have TypeScript correctly derive the
more coerced type.