style(monatsplaner): improve terminology around gross and net
There is some confusion about which monetary values are gross and which
are net. Furthermore there is a collision in terminology of "Elterngeld
(netto)" and "Elterngeldnetto".
The included changes are a first effort to help on this problem with the
help of content design resources. Further iterations are expected.

Minimal boyscout work.