feat(queries): improve queries for Ihr Nachwuchs
Primarily addresses the issues that questions were asked like the child
wasn't yet born. But further improvements and hints were added too.
The future structure is not yet fully refined. But this is a step into
the right direction.

Noticeably side effect also for the end-to-end tests is that per default
a single child is no expected, instead of the zero before. Because many
tests use the increase/decrease buttons.