feat(metabase): implement funnel metrics and replace table secret with variable
This update introduces funnel metrics to enhance tracking.
The decision to replace the table secret with a variable
leans towards convenience over security, as the tableId
functions primarily as a different representation of the
table name, which isn't inherently sensitive. Restricting
access to tableIds as secrets would require changes across
multiple areas of the app, especially when adding new tables.
Therefore, using a variable for tableId provides a more
flexible and manageable solution without compromising security.