fix(monatsplaner): modifier combos in input for Bruttoeinkommen
The input component for the Bruttoeinkommen filters key events to
prevent the user to input non numeric values. This is for user
experience and validation simplicity.
In return, it is somewhat hard not to filter too much. There was already
a collection of "special" keys (like backspace or arrows) to accept
anyway. This no got extended to accept any kind of combos using modifier
keys like the control key. It allows users to use system shortcuts like
select-all or copy-paste.
Unfortunately this behavior can't be properly tested.