fix(CI): vulnerability scan pipeline flakiness
Trivy is known to have issues with is database connection and caching on
old versions. This led to many failed scans in a flaky manner.
More recent versions are known to fix this issue. Potentially it will be
necessary to configure additional database connections in future if it
still remains too flaky.

Furthermore the workflow has been simplified by making use of the Trivy
actions own capabilities to fail the pipeline if vulnerabilities of
matching severity are found.
Notice that we never uploaded the output file as we still don't have
access to GitHub its security tab for private repositories.