test: use style-sheets for component testing
Some styles have an effect on the unit tests. These are primarily those
which affect the visibility of elements. These are picked up by the
testing-library we are using. But this did not work for the utility
classes we use from TailwindCSS its atomic styling. To make these CSS
classes be "understood" by the test setup, they must be enabled and
imported.

This allowed us to write some tests which were not possible before
(some were marked as TODOs). But it also showed us tests which just
worked out by "luck" so far.