test: fix test for the hook to calculate Elterngeld
Tests where green even not correct. Fortunately, the code under test is
actually correct. The broken assertion was hiding another test flaw that
could be fixed.
Checkout the comments in code for details. The final solution does not
feel optimal. With future refactorings of the data structures for the
algorithm, it might improve and gets revisited to become more intuitive.

In result we can rely on this test suite again and continue adding
features.