chore(calculations): add constructor for ErwerbsZeitraumLebensMonat
This seem to be a non optimal practice from the imported Java code.
Though, even in Java it would be arguably much better to use
a constructor. Anyways, code got easier and more terse. Absolutely minor
improvements to surrounding code to don't break anything that is not
properly covered by tests.