fix(calculations): filter Einnkommen for Lebensmonate ohne Variante
The calculation algorithm has a few bugs. One of them is in the area
that Einkommen in the Bezugszeitraum is not properly separated and
filtered. For example we do know about a bug where Einkommen is not
grouped by Basiselterngeld and ElterngeldPlus as the law requires.

The planer introduced the possibility to define Einkommen for any month,
with or without a chosen Variante. This flushed up a new symptom of the
bug where these Einkommen were actually took into calculation for
Elterngeldbezüge. Which is very obviously wrong and confusing to the
user.

To cover this bug quickly, the connection between the planer and
calculation algorithm filters such Einkommen. This does not solve the
original bug within the calculation.
Though, it can still count as viable solution as this is data the
calculation simply don't need and would have to filter itself. So
"pre-filtering" it is not an issue. Though, it should become unnecessary
once the algorithm correctly handles Einkommen.