style: use intrinsic styling without media queries for responsive split view
The split view used strict breaking points based on the width of the
screen. This is not the best design approach. Using more intrinsic
styling, relying on the browsers ability to distribute space based on
the actual content and available space is more stable and reliable.

The desired split view feature is actually a quite trivial flex layout
with wrapping. Plus some minor extras. The usage got actually much
simpler and became more flexible.
Though, this removes a "feature" that can not easily be implemented with
this approach. This is the conditional styling of the background color
for the split items. Arguably, this is no unified design and should be
seriously questioned why necessary. Due to the proper usage of sections
and headings, it is actually not necessary and can be removed.