style: remove a bunch of simple but needless media queries
Some of them do minor tweaks for spacing or font size. Which is actually
not necessary. A unified styling that is adaptive works out better long
term and provides more stable results.

The adaptive font root font sizes where planned to be removed anyway.
They are not part of our styling theme and don't bring any value for the
user.

For some unknown reason there were some awkward issues in the end-to-end
tests. The screenshot diff where expected. But some reason there was an
old baseline for the long ago removed `(netto)` annotations on the
summary page.
Moreover, one test double clicked the `Einkünfte aus nicht selständiger
Arbeit` which was disabling it again. Confusing.