test(end-to-end): improve coverage of screenshot tests
We have many end-to-end test case scenario that compare multiple
screenshots. When the first screenshot does not match, the whole test
case fails. If the developer acknowledges the change and use
`--update-snapshots` for the next test run, it will accept and update
also any screenshot that comes afterwards. So there might be changes
which are no more detected by the baseline.
To solve this issue, we use now "soft" assertions by Playwright. They
make sure that the test case continues running and only fails at the end
(or when asked to). So the developer is faced with ALL not matching
screenshots and can verify them.

The same approach is used for the accessibility checks. Though, here the
issue is not the same. But it helps to fix accessibility issues faster,
as all will be reported together, instead of fail, fix, run, new fail
later in test case, fix run, ...